+++ back/babble/src/main/java/gg/babble/babble/DataLoader.java
@@ -1,6 +1,9 @@
 package gg.babble.babble;
 
 import gg.babble.babble.domain.*;
+import gg.babble.babble.domain.repository.GameRepository;
+import gg.babble.babble.domain.repository.TagRepository;
+import gg.babble.babble.domain.repository.UserRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
