+++ back/babble/src/main/java/gg/babble/babble/domain/repository/RoomRepository.java
@@ -1,5 +1,6 @@
-package gg.babble.babble.domain;
+package gg.babble.babble.domain.repository;
 
+import gg.babble.babble.domain.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomRepository extends JpaRepository<Room, Long> {
