+++ back/babble/src/main/java/gg/babble/babble/service/TagService.java
@@ -1,7 +1,7 @@
 package gg.babble.babble.service;
 
 import gg.babble.babble.domain.Tag;
-import gg.babble.babble.domain.TagRepository;
+import gg.babble.babble.domain.repository.TagRepository;
 import gg.babble.babble.exception.BabbleNotFoundException;
 import org.springframework.stereotype.Service;
