+++ back/babble/src/main/java/gg/babble/babble/service/UserService.java
@@ -1,7 +1,7 @@
 package gg.babble.babble.service;
 
 import gg.babble.babble.domain.User;
-import gg.babble.babble.domain.UserRepository;
+import gg.babble.babble.domain.repository.UserRepository;
 import gg.babble.babble.exception.BabbleNotFoundException;
 import org.springframework.stereotype.Service;
