+++ back/babble/src/test/java/gg/babble/babble/domain/RoomRepositoryTest.java
@@ -2,16 +2,15 @@ package gg.babble.babble.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import gg.babble.babble.domain.repository.RoomRepository;
 import gg.babble.babble.service.GameService;
 import gg.babble.babble.service.TagService;
 import gg.babble.babble.service.UserService;
 import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.Arrays;
