+++ back/babble/src/test/java/gg/babble/babble/domain/TagRepositoryTest.java
@@ -2,6 +2,7 @@ package gg.babble.babble.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import gg.babble.babble.domain.repository.TagRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
